+++ network/data/src/main/kotlin/me/proton/core/network/data/protonApi/BaseRetrofitApi.kt
@@ -19,7 +19,6 @@ package me.proton.core.network.data.protonApi
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import me.proton.core.network.data.humanverification.VerificationMethodApi
 import me.proton.core.network.data.mapper.parseDetails
 import me.proton.core.network.domain.ApiResult
 import retrofit2.http.Body
@@ -55,7 +54,7 @@ data class ProtonErrorData(
 @Serializable
 data class Details(
     @SerialName("HumanVerificationMethods")
-    val verificationMethods: List<VerificationMethodApi>? = null,
+    val verificationMethods: List<String>? = null,
     @SerialName("HumanVerificationToken")
     val verificationToken: String? = null
 )