+++ network/domain/src/main/kotlin/me/proton/core/network/domain/humanverification/HumanVerificationDetails.kt
@@ -22,7 +22,7 @@ import me.proton.core.network.domain.client.ClientId
 
 data class HumanVerificationDetails(
     val clientId: ClientId,
-    val verificationMethods: List<VerificationMethod>,
+    val verificationMethods: List<String>,
     val verificationToken: String? = null,
     val state: HumanVerificationState,
     val tokenType: String? = null,